Se divide responsabilidades de la clase para la capa de data.
